Add files via upload
Stephen Fahy
G00172772
WEB SITE DESIGN
13/12/2015

The projectmain file contains all the files for my South Dublin Housing Council website. 
This consists of my six html5 webpages. 
There is a homepage, a tenants page, a property page, a repairs page, a tradesman page and an equipment page.
Each webpage is linked to each other by a sidebar on every webpage with links to the other pages. 
You enter the website via the homepage which then can access all the other webpages of the website. 
In the projectmain file there is the css file whch contains the external css design for the website, 
a images file which contains the images for my website, a js file which contains the javascript for the website, 
and a php file which contains the php insert statements and select statements for each webpage except the homepage.

The website is linked to a database via HTML5 forms and my php code on each webpage except the homepage. 
The database can be accessed by opening the south_dublin